Move css-errors out of stores as it is no longer a store
