+++ projects/plugins/boost/app/assets/src/js/features/critical-css/cloud-css-meta/cloud-css-meta.tsx
@@ -8,7 +8,7 @@ import {
 	useCriticalCssState,
 	useRegenerateCriticalCssAction,
 } from '../lib/stores/critical-css-state';
-import { getCriticalCssIssues } from '../lib/stores/critical-css-state-errors';
+import { getCriticalCssIssues } from '../lib/critical-css-errors';
 
 type CloudCssMetaProps = {
 	isCloudCssAvailable: boolean;
