+++ projects/plugins/boost/app/assets/src/js/features/critical-css/index.ts
@@ -2,4 +2,4 @@ export { RegenerateCriticalCssSuggestion } from './regenerate-critical-css-sugge
 export { criticalCssState } from './lib/stores/critical-css-state';
 export { suggestRegenerateDS, type RegenerationReason } from './lib/stores/suggest-regenerate';
 export { type Provider } from './lib/stores/critical-css-state-types';
-export { groupErrorsByFrequency } from './lib/stores/critical-css-state-errors';
+export { groupErrorsByFrequency } from './lib/critical-css-errors';
