+++ projects/plugins/boost/app/assets/src/js/features/critical-css/lib/describe-critical-css-recommendations.ts
@@ -1,7 +1,7 @@
 import { __, _n, sprintf } from '@wordpress/i18n';
 import { CriticalCssErrorDetails } from './stores/critical-css-state-types';
 import { castToNumber } from '$lib/utils/cast-to-number';
-import type { ErrorSet } from './stores/critical-css-state-errors';
+import type { ErrorSet } from './critical-css-errors';
 import type { ComponentType } from 'react';
 import UrlComponentsExample from '$features/critical-css/url-components-example/url-components-example';
